feat: Article과 User의 관계 매핑 및 Interceptor 적용

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -5,6 +5,7 @@ import techcourse.myblog.article.Article;
 import techcourse.myblog.article.ArticleRepository;
 import techcourse.myblog.exception.NotFoundObjectException;
 import techcourse.myblog.service.dto.ArticleDto;
+import techcourse.myblog.user.User;
 
 import javax.transaction.Transactional;
 
@@ -16,8 +17,9 @@ public class ArticleService {
         this.articleRepository = articleRepository;
     }
 
-    public Article createArticle(ArticleDto articleDto) {
-        Article article = articleDto.toEntity();
+    public Article createArticle(ArticleDto articleDto, User author) {
+        Article article = articleDto.toEntity(author);
+
         return articleRepository.save(article);
     }
 
@@ -29,12 +31,15 @@ public class ArticleService {
     @Transactional
     public Article updateArticle(Long articleId, Article updatedArticle) {
         Article article = findArticle(articleId);
+        article.checkCorrespondingAuthor(updatedArticle.getAuthor());
         article.update(updatedArticle);
 
         return article;
     }
 
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, User user) {
+        Article article = findArticle(articleId);
+        article.checkCorrespondingAuthor(user);
         articleRepository.deleteById(articleId);
     }
 }
